move event_emitter_mixin to shell/browser
