+++ filenames.gni
@@ -177,6 +177,8 @@ filenames = {
     "shell/browser/electron_speech_recognition_manager_delegate.h",
     "shell/browser/electron_web_ui_controller_factory.cc",
     "shell/browser/electron_web_ui_controller_factory.h",
+    "shell/browser/event_emitter_mixin.cc",
+    "shell/browser/event_emitter_mixin.h",
     "shell/browser/feature_list.cc",
     "shell/browser/feature_list.h",
     "shell/browser/font_defaults.cc",
@@ -515,8 +517,6 @@ filenames = {
     "shell/common/gin_helper/event_emitter.h",
     "shell/common/gin_helper/event_emitter_caller.cc",
     "shell/common/gin_helper/event_emitter_caller.h",
-    "shell/common/gin_helper/event_emitter_mixin.cc",
-    "shell/common/gin_helper/event_emitter_mixin.h",
     "shell/common/gin_helper/function_template.cc",
     "shell/common/gin_helper/function_template.h",
     "shell/common/gin_helper/function_template_extensions.h",
